update: cancel batch event handling post process
